Remove prototype CUDA kernels from package
The batch Kalman filter was using two prototype custom CUDA kernels for performing fast Cholesky and regular-transposed matrix multiplication. The CUBLAS version of the former (technically LU) has been sped up now by better pointer creation in CUDA.jl. The second is still very slow and is the bottleneck to the computation. The prototype replacement is not fully tested though and even though it passed this packages unit tests, I found a case where it was giving unexplained incorrect values. For safety, it will be removed for now until the custom batch kernels are fully tested.