Make Kalman update compatible with SArrays and Zygote.
hermitianpart is not compatible with StaticArrays and results in a Matrix output. I will make a PR to StaticArrays.jl that implements an optimised implementation, but for now we can use this.

We perform the cholesky decomposition on S so we use that factorisation for the solve and logpdf, which will be more numerically stable and avoid redundant computations.

There are some issues with Zygote and PDMat (https://github.com/JuliaStats/PDMats.jl/issues/159) which don't appear to be resolved, so we use the Cholesky directly for the solve before wrapping in PDMat, which is required for MvNormal.