Use explicit type parameter for resampling state construction.
For some reason, the previous `map` used for resampling was causing type
inference to fail, with the return type of `resample` being
`GeneralisedFilters.ParticleDistribution{Float64, PT} where PT<:GeneralisedFilters.Particle`.
It's not entirely clear why this was happening, but changing to an explicit
type parameter seems to have resolved the issue.